workers/deteccao/requirements.txt
opencv-python>=4.7.0.72
pika>=1.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
mediapipe>=0.10.0
pymongo>=4.5.0
minio>=7.1.1
protobuf>=3.20.0,<4.0.0
